Splits formknoppen.class.php
