+++ lib/view/formulier/Formulier.class.php
@@ -11,7 +11,9 @@ require_once 'view/formulier/GetalVelden.class.php';
 require_once 'view/formulier/KeuzeVelden.class.php';
 require_once 'view/formulier/BBCodeVelden.class.php';
 require_once 'view/formulier/UploadVelden.class.php';
-require_once 'view/formulier/FormKnoppen.class.php';
+require_once 'view/formulier/knoppen/FormKnoppen.class.php';
+require_once 'view/formulier/knoppen/FormDefaultKnoppen.class.php';
+require_once 'view/formulier/knoppen/ModalCloseButtons.class.php';
 
 /**
  * Formulier.class.php
