+++ lib/view/formulier/knoppen/DeleteKnop.class.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * DeleteKnop.class.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ */
+class DeleteKnop extends FormulierKnop {
+
+	public function __construct($url, $action = 'post confirm ReloadPage', $label = 'Verwijderen', $title = 'Definitief verwijderen', $icon = 'cross') {
+		parent::__construct($url, $action, $label, $title, $icon);
+	}
+
+}
