+++ lib/view/formulier/knoppen/ResetKnop.class.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * ResetKnop.class.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ */
+class ResetKnop extends FormulierKnop {
+
+	public function __construct($url = null, $action = 'reset', $label = 'Reset', $title = 'Reset naar opgeslagen gegevens', $icon = 'arrow_rotate_anticlockwise') {
+		parent::__construct($url, $action, $label, $title, $icon);
+	}
+
+}
