+++ lib/view/formulier/knoppen/SubmitKnop.class.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * SubmitKnop.class.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ */
+class SubmitKnop extends FormulierKnop {
+
+	public function __construct($url = null, $action = 'submit', $label = 'Opslaan', $title = 'Invoer opslaan', $icon = 'disk') {
+		parent::__construct($url, $action, $label, $title, $icon);
+	}
+
+}
